Move and refactor e2e test
Ref oursky/likecoin-chain#148
